requirements.txt
python-dotenv
streamlit
azure-core
azure-storage-blob
azure-ai-documentintelligence
